+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add self update with `uamp update`.
 - Support winamp way of looking up images.
 - Add option to print aliases with `uamp cfg --aliases`.
+- Add verbosity option. For now works only for `cfg --aliases`.
 
 ### Changes
 - Playlist add will now use playlist add policy of all playlists in the stack.